+++ samples/concepts.js
@@ -39,10 +39,12 @@ let datastore = {
   save: makeStub(),
 };
 
+const namespace = `${Date.now()}`;
 class TestHelper {
   constructor(projectId) {
     const options = {
       projectId: projectId,
+      namespace,
     };
     this.datastore = new Datastore(options);
   }
@@ -53,6 +55,7 @@ class Entity extends TestHelper {
     super(projectId);
     // To create the keys, we have to use this instance of Datastore.
     datastore.key = this.datastore.key;
+    datastore.namespace = this.datastore.namespace;
 
     this.incompleteKey = this.getIncompleteKey();
     this.namedKey = this.getNamedKey();
@@ -981,7 +984,7 @@ async function transferFunds(fromKey, toKey, amount) {
     },
   ]);
 
-  return transaction.commit();
+  return await transaction.commit();
 }
 // [END datastore_transactional_update]
 
@@ -1106,7 +1109,7 @@ class Transaction extends TestHelper {
         } else {
           // Create the task entity.
           transaction.save(taskEntity);
-          transaction.commit();
+          await transaction.commit();
         }
         return taskEntity;
       } catch (err) {